% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_geographies}
\alias{get_geographies}
\title{Get Available Geographies or Region List}
\usage{
get_geographies(geography = NULL)
}
\arguments{
\item{geography}{Character. Spatial level code (e.g. "KRE"). If NULL, returns all levels.}
}
\value{
A data frame with \code{ID} and \code{Name}.
}
\description{
Retrieves a list of available spatial levels (if \code{geography} is NULL) or
a list of regions for a specific level (e.g., "KRE").
}
