% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{select_years}
\alias{select_years}
\title{Interactively Select Years}
\usage{
select_years(variable, level)
}
\arguments{
\item{variable}{Indicator ID.}

\item{level}{Spatial level ID.}
}
\value{
Character vector of years.
}
\description{
Probes the API for available years for a specific indicator and level,
then allows the user to select one or more years.
}
