% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_inkaR}
\alias{theme_inkaR}
\title{Premium ggplot2 theme for inkaR}
\usage{
theme_inkaR(mode = c("light", "dark"), base_size = 11)
}
\arguments{
\item{mode}{Character. "light" or "dark".}

\item{base_size}{Numeric. Base font size.}
}
\description{
Premium ggplot2 theme for inkaR
}
