% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{search_indicators}
\alias{search_indicators}
\title{Search Indicators}
\usage{
search_indicators(pattern, lang = c("de", "en"))
}
\arguments{
\item{pattern}{Regex pattern to search for.}

\item{lang}{Language to search in ("de" or "en").}
}
\value{
A filtered tibble of indicators.
}
\description{
Search for indicators by name or description.
}
